+++ app/src/main/java/rees46/demo_android/features/cardProduct/CardProductView.kt
@@ -1,4 +1,4 @@
-package rees46.demo_android.core_ui
+package rees46.demo_android.features.cardProduct
 
 import android.content.Context
 import android.util.AttributeSet
@@ -6,7 +6,7 @@ import android.widget.ImageButton
 import android.widget.TextView
 import com.google.android.material.button.MaterialButton
 import rees46.demo_android.R
-import rees46.demo_android.base.BaseCardProductView
+import rees46.demo_android.core_ui.BaseCardProductView
 import rees46.demo_android.features.product.Product
 
 class CardProductView @JvmOverloads constructor(
