+++ app/src/main/java/rees46/demo_android/features/main/category/CardProductView.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.util.AttributeSet
 import com.google.android.material.button.MaterialButton
 import rees46.demo_android.R
-import rees46.demo_android.base.BaseCardProductView
+import rees46.demo_android.core_ui.BaseCardProductView
 
 class CardProductView @JvmOverloads constructor(
     context: Context,
