+++ app/src/main/java/rees46/demo_android/features/main/home/HomeFragment.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.launch
 import org.koin.androidx.viewmodel.ext.android.viewModel
 import rees46.demo_android.core.view.BaseFragment
 import rees46.demo_android.databinding.FragmentHomeBinding
-import rees46.demo_android.core_ui.RecommendationBlockView
+import rees46.demo_android.features.recommendationBlock.RecommendationBlockView
 import rees46.demo_android.features.product.Product
 
 class HomeFragment
