+++ app/src/main/java/rees46/demo_android/features/recommendationBlock/CardProductsAdapter.kt
@@ -1,39 +1,37 @@
 package rees46.demo_android.features.recommendationBlock
 
 import android.content.Context
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import rees46.demo_android.core_ui.ShortCardProductView
 import rees46.demo_android.features.product.Product
 
-class ShortCardProductsAdapter(
+class CardProductsAdapter(
     private val context: Context,
     private val products: List<Product>,
     private val cardProductViewSettings: CardProductViewSettings,
     private val listener: ClickListener
-) : RecyclerView.Adapter<ShortCardProductsAdapter.ViewHolder>() {
+) : RecyclerView.Adapter<CardProductsAdapter.ViewHolder>() {
 
     interface ClickListener {
         fun onCardProductClick(product: Product)
     }
 
-    inner class ViewHolder(private val view: View, private val listener: ClickListener)
-        : RecyclerView.ViewHolder(view) {
+    inner class ViewHolder(private val cardProductView: CardProductView, private val listener: ClickListener)
+        : RecyclerView.ViewHolder(cardProductView) {
 
         fun bind(product: Product) {
-            (view as ShortCardProductView).updateProduct(product)
+            cardProductView.updateProduct(product)
 
-            view.setOnClickListener {
+            cardProductView.setOnClickListener {
                 listener.onCardProductClick(product)
             }
         }
     }
 
     override fun onCreateViewHolder(viewGroup: ViewGroup, viewType: Int): ViewHolder {
-        val shortCardProductView = ShortCardProductView(context, cardProductViewSettings,null)
+        val cardProductView = CardProductView(context, cardProductViewSettings,null)
 
-        return ViewHolder(shortCardProductView, listener)
+        return ViewHolder(cardProductView, listener)
     }
 
     override fun onBindViewHolder(viewHolder: ViewHolder, position: Int) {
