+++ app/src/main/res/layout/fragment_card_product.xml
@@ -10,7 +10,7 @@
         android:layout_height="match_parent"
         android:layout_width="match_parent">
 
-        <rees46.demo_android.core_ui.CardProductView
+        <rees46.demo_android.features.cardProduct.CardProductView
             android:id="@+id/card_product_view"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
@@ -18,7 +18,7 @@
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
 
-        <rees46.demo_android.core_ui.RecommendationBlockView
+        <rees46.demo_android.features.recommendationBlock.RecommendationBlockView
             android:id="@+id/recommendation_block"
             android:layout_width="0dp"
             android:layout_height="wrap_content"
