+++ app/src/main/res/layout/view_recommendation_block.xml
@@ -40,5 +40,5 @@
         app:layout_constraintTop_toBottomOf="@+id/header_text"
         android:orientation="horizontal"
         app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        tools:listitem="@layout/view_short_card_product" />
+        tools:listitem="@layout/view_recommendation_card_product" />
 </androidx.constraintlayout.widget.ConstraintLayout>
